requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
currencyapicom==0.1.1
Django==4.2.3
django-apscheduler==0.6.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-redis==5.3.0
djangorestframework==3.14.0
everapi==0.1.1
idna==3.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.3
